Drive-by #2142: retire the legacy incremental merge SSE events; merge_state is the sole live channel
merge_progress, merge_queued, and merge_queue_updated had zero consumers
since #2129 but still cost two frames per ffmpeg tick and one DB read per
queued book per promotion/cancel. Removed end-to-end: emit sites,
emitQueueEvent + emitQueuePositionUpdates, schema entries, SSEEventPayloads
keys, SSE_PARSERS wirings, cache-matrix rows. merge_started (toast +
event-history) and all terminals remain; the useEventSource comment now
names merge_started as the sole surviving incremental.

Riders, same pass:
- merge-minimum shared via listTopLevelAudioFiles/requireMergeMinimum
  (third inline copy in executeMerge unified with the eligibility gates)
- executeMerge pre-flight early returns converted to MergeError throws —
  vanished-book/vanished-ffmpeg now report merge_failed instead of a
  silently disappearing chip
- titleFor owns its Book-N fallback; updateProgress broadcasts only when
  the book is tracked; pruneAfterWriteFailure shared by emitTo/writeToAll
- hold-branch log line folded into recordSearchRelaxedHeldEvent (one home,
  structured attempt field distinguishes the retry path) + helper tests
- Pick<SettingsService,'get'> symmetry in validateBookForMerge
- SSE event-count pin deduped; stale AC24 comment reworded
- countOccurrences self-overlap claim pinned (conservative count + hold)
- #2129 snapshot tests relocated to merge-state.test.ts (shared fixtures
  in __tests__/merge-fixtures.ts); badge-render assertions live only in
  the badge suite

Transition caveat (accepted, pre-1.0): a tab on a pre-rollup bundle loses
live merge progress until refresh.
