+++ src/client/lib/sse/safe-parse-event.ts
@@ -13,10 +13,7 @@ import {
   reviewNeededPayload,
   mergeCompletePayload,
   mergeStartedPayload,
-  mergeProgressPayload,
   mergeFailedPayload,
-  mergeQueuedPayload,
-  mergeQueueUpdatedPayload,
   mergeStatePayload,
   searchStartedPayload,
   searchIndexerCompletePayload,
@@ -59,10 +56,7 @@ export const SSE_PARSERS: SSEParserMap = {
   review_needed: reviewNeededPayload,
   merge_complete: mergeCompletePayload,
   merge_started: mergeStartedPayload,
-  merge_progress: mergeProgressPayload,
   merge_failed: mergeFailedPayload,
-  merge_queued: mergeQueuedPayload,
-  merge_queue_updated: mergeQueueUpdatedPayload,
   merge_state: mergeStatePayload,
   search_started: searchStartedPayload,
   search_indexer_complete: searchIndexerCompletePayload,
